requirements.txt
# Core detection (YOLOv11)
ultralytics>=8.3.0

# Vision & ML
opencv-python>=4.13.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0

# Data & plotting
numpy>=1.23.0
pandas>=2.2.3
matplotlib>=3.9.4
seaborn>=0.13.2
scipy>=1.13.1

# Utilities
tqdm>=4.67.3
PyYAML>=6.0.3
